First refactor the manage view a bunch
